+++ src/aptapy/models.py
@@ -17,7 +17,7 @@
 """
 
 from numbers import Number
-from typing import Tuple
+from typing import Tuple, Union
 
 import matplotlib
 import numpy as np
@@ -31,6 +31,7 @@ from .modeling import (
     AbstractFitModel,
     AbstractSigmoidFitModel,
     FitParameter,
+    FitStatus,
     PhonyCRVFitModel,
     wrap_rv_continuous,
 )
@@ -547,6 +548,59 @@ class Gaussian(AbstractFitModel):
         self.amplitude.init(scipy.integrate.trapezoid(ydata, xdata))
         self.mu.init(np.average(xdata, weights=ydata))
         self.sigma.init(np.sqrt(np.average((xdata - self.mu.value)**2, weights=ydata)))
+
+    def fit_iterative(self, xdata: Union[ArrayLike, Histogram1d], ydata: ArrayLike = None, *,
+            p0: ArrayLike = None, sigma: ArrayLike = None, num_sigma_left: float = 2.,
+            num_sigma_right: float = 2., num_iterations: int = 2, **kwargs) -> "FitStatus":
+        """Fit the core of Gaussian data within a given number of sigma around the peak.
+
+        This function performs a first round of fit to the data (either a histogram or
+        scatter plot data) and then repeats the fit iteratively, limiting the fit range
+        to a specified interval defined in terms of deviations (in sigma) around the peak.
+
+        Arguments
+        ----------
+        xdata : array_like or Histogram1d
+            The data (scatter plot x values) or histogram to fit.
+
+        ydata : array_like, optional
+            The y data to fit (if xdata is not a Histogram1d).
+
+        p0 : array_like, optional
+            The initial values for the fit parameters.
+
+        sigma : array_like, optional
+            The uncertainties on the y data.
+
+        num_sigma_left : float
+            The number of sigma on the left of the peak to be used to define the
+            fitting range.
+
+        num_sigma_right : float
+            The number of sigma on the right of the peak to be used to define the
+            fitting range.
+
+        num_iterations : int
+            The number of iterations of the fit.
+
+        kwargs : dict, optional
+            Additional keyword arguments passed to `fit()`.
+
+        Returns
+        -------
+        FitStatus
+            The results of the fit.
+        """
+        fit_status = self.fit(xdata, ydata, p0=p0, sigma=sigma, **kwargs)
+        for i in range(num_iterations):
+            xmin = self.mean() - num_sigma_left * self.std()
+            xmax = self.mean() + num_sigma_right * self.std()
+            try:
+                fit_status = self.fit(xdata, ydata, p0=self.parameter_values(),
+                                      sigma=sigma, xmin=xmin, xmax=xmax, **kwargs)
+            except RuntimeError as exception:
+                raise RuntimeError(f"Exception after {i+1} iteration(s)") from exception
+        return fit_status
 
     def default_plotting_range(self) -> Tuple[float, float]:
         """Overloaded method.