+++ tests/test_modeling.py
@@ -104,7 +104,7 @@ def test_gaussian_fit():
     plt.figure(inspect.currentframe().f_code.co_name)
     model = Gaussian()
     TEST_HISTOGRAM.plot()
-    model.fit_histogram(TEST_HISTOGRAM)
+    model.fit(TEST_HISTOGRAM)
     model.plot(fit_output=True)
     assert model.mu.compatible_with(0., NUM_SIGMA)
     assert model.sigma.compatible_with(1., NUM_SIGMA)
@@ -118,7 +118,7 @@ def test_gaussian_fit_subrange():
     plt.figure(inspect.currentframe().f_code.co_name)
     model = Gaussian()
     TEST_HISTOGRAM.plot()
-    model.fit_histogram(TEST_HISTOGRAM, xmin=-2., xmax=2.)
+    model.fit(TEST_HISTOGRAM, xmin=-2., xmax=2.)
     model.plot(fit_output=True)
     assert model.mu.compatible_with(0., NUM_SIGMA)
     assert model.sigma.compatible_with(1., NUM_SIGMA)
@@ -134,7 +134,7 @@ def test_gaussian_fit_bound():
     model.mu.minimum = 0.05
     model.mu.value = 0.1
     TEST_HISTOGRAM.plot()
-    model.fit_histogram(TEST_HISTOGRAM)
+    model.fit(TEST_HISTOGRAM)
     model.plot(fit_output=True)
     assert model.mu.value >= model.mu.minimum
     plt.legend()
@@ -148,7 +148,7 @@ def test_gaussian_fit_frozen():
     # Calculate the normalization from the histogram.
     model.amplitude.freeze(TEST_HISTOGRAM.area())
     TEST_HISTOGRAM.plot()
-    model.fit_histogram(TEST_HISTOGRAM)
+    model.fit(TEST_HISTOGRAM)
     model.plot(fit_output=True)
     assert model.mu.compatible_with(0., NUM_SIGMA)
     assert model.sigma.compatible_with(1., NUM_SIGMA)
@@ -166,11 +166,43 @@ def test_gaussian_fit_frozen_and_bound():
     model.mu.minimum = 0.05
     model.mu.value = 0.1
     TEST_HISTOGRAM.plot()
-    model.fit_histogram(TEST_HISTOGRAM)
+    model.fit(TEST_HISTOGRAM)
     model.plot(fit_output=True)
     assert model.mu.value >= model.mu.minimum
     assert model.sigma.value == 1.1
     plt.legend()
+
+
+def test_gaussian_fit_iterative_hist():
+    """Test an iterative fit for a Gaussian histogram.
+    """
+    plt.figure(inspect.currentframe().f_code.co_name)
+    model = Gaussian()
+    print(model)
+    TEST_HISTOGRAM.plot()
+    model.fit_iterative(TEST_HISTOGRAM, num_iterations=3, num_sigma_left=3., num_sigma_right=3.)
+    model.plot(fit_output=True)
+    plt.legend()
+    assert model.mu.compatible_with(0., NUM_SIGMA)
+    assert model.sigma.compatible_with(1., NUM_SIGMA)
+    assert model.status.pvalue > 0.001
+
+
+def test_gaussian_fit_iterative_scatter():
+    """Test an iterative fit for a Gaussian scatter plot.
+    """
+    plt.figure(inspect.currentframe().f_code.co_name)
+    model = Gaussian(xlabel="x [a.u.]", ylabel="y [a.u.]")
+    print(model)
+    sigma = 0.01
+    xdata, ydata = model.random_fit_dataset(sigma, seed=313)
+    plt.errorbar(xdata, ydata, sigma, fmt="o", label="Random data")
+    model.fit_iterative(xdata, ydata, sigma=sigma, num_iterations=3)
+    model.plot(fit_output=True)
+    plt.legend()
+    assert model.mu.compatible_with(0., NUM_SIGMA)
+    assert model.sigma.compatible_with(1., NUM_SIGMA)
+    assert model.status.pvalue > 0.001
 
 
 def test_sum_gauss_line():
@@ -183,7 +215,7 @@ def test_sum_gauss_line():
     hist.fill(x)
     model = Gaussian() + Line()
     hist.plot()
-    model.fit_histogram(hist)
+    model.fit(hist)
     model.plot(fit_output=True)
     plt.legend()